feat(response): deal with mime types
Now, we'll reply with the correct mime type depending on the
request file. For this, we'll be using the mime module, and whenever
it fails, we'll use the mime type coming from GitHub, which usually
is text/plain or application/octet-stream, so we don't have to read the
file on our side and try to identify if it's text or binary.

Closes #2